# openapi-mvp-v2.yaml
openapi: 3.0.3
info:
  title: Nomad API (MVP v2)
  version: 2.0.0
  description: |
    Expanded MVP API contract for Nomad with full hook coverage:
    - Auth + Onboarding restored
    - CRUD-expanded resources for courses, communities, posts, and messaging
servers:
  - url: http://localhost:3000/api
    description: Local dev
  - url: https://api.nomad.example.com
    description: Production (placeholder)
...
security:
  - bearerAuth: []
paths:
  /me:
    get:
      tags: [Users]
      summary: Get current user
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      tags: [Users]
      summary: Update current user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      tags: [Users]
      summary: Delete current user account
      responses:
        "204":
          description: Account removed
  /users:
    get:
      tags: [Users]
      summary: List/search users
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: university
          schema: { type: string }
        - in: query
          name: major
          schema: { type: string }
        - in: query
          name: year
          schema: { type: string }
        - in: query
          name: connected
          schema: { type: boolean }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedUsers" }
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user profile
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{userId}/connect:
    post:
      tags: [Users]
      summary: Connect with user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionStatusResponse" }
    delete:
      tags: [Users]
      summary: Remove connection
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionStatusResponse" }
  /courses:
    get:
      tags: [Courses]
      summary: List courses
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: department
          schema: { type: string }
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [Intro, Intermediate, Advanced, Graduate]
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCourses" }
    post:
      tags: [Courses]
      summary: Create course
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCourseRequest" }
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
...
  /communities:
    get:
      tags: [Communities]
      summary: List communities
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: joined
          schema: { type: boolean }
      responses:
        "200":
          description: Communities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedCommunities" }
    post:
      tags: [Communities]
      summary: Create community
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommunityRequest" }
      responses:
        "201":
          description: Created community
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Community" }
...
  /messages/conversations:
    get:
      tags: [Messages]
      summary: List conversations
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedConversations" }
    post:
      tags: [Messages]
      summary: Create conversation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConversationRequest" }
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
  /messages/conversations/{conversationId}:
    patch:
      tags: [Messages]
      summary: Update conversation metadata
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateConversationRequest" }
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
    delete:
      tags: [Messages]
      summary: Delete conversation
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "204":
          description: Conversation deleted
  /messages/conversations/{conversationId}/messages:
    get:
      tags: [Messages]
      summary: List messages
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedMessages" }
    post:
      tags: [Messages]
      summary: Send message
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMessageRequest" }
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
  /messages/conversations/{conversationId}/messages/{messageId}:
    delete:
      tags: [Messages]
      summary: Delete message
      parameters:
        - $ref: "#/components/parameters/ConversationId"
        - in: path
          name: messageId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Message deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: string }
...
    ConversationId:
      in: path
      name: conversationId
      required: true
      schema: { type: string }
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        avatar: { type: string }
        university: { type: string }
        major: { type: string }
        minor: { type: string, nullable: true }
        year: { type: string }
        bio: { type: string }
        headline: { type: string, nullable: true }
        interests:
          type: array
          items: { type: string }
        courses:
          type: array
          items: { type: string }
        communities:
          type: array
          items: { type: string }
        isConnected: { type: boolean }
        isOnline: { type: boolean }
    UpdateUserRequest:
      type: object
      properties:
        name: { type: string }
        bio: { type: string }
        headline: { type: string }
        major: { type: string }
        minor: { type: string }
        year: { type: string }
        interests:
          type: array
          items: { type: string }
    Course:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        title: { type: string }
        credits: { type: integer }
        description: { type: string }
        prerequisites:
          type: array
          items: { type: string }
        nextCourses:
          type: array
          items: { type: string }
        department: { type: string }
        difficulty:
          type: string
          enum: [Intro, Intermediate, Advanced, Graduate]
        tags:
          type: array
          items: { type: string }
        students:
          type: array
          items: { type: string }
        rating: { type: number, format: float }
    CreateCourseRequest:
      type: object
      properties:
        code: { type: string }
        title: { type: string }
        credits: { type: integer }
        description: { type: string }
        prerequisites:
          type: array
          items: { type: string }
        nextCourses:
          type: array
          items: { type: string }
        department: { type: string }
        difficulty:
          type: string
          enum: [Intro, Intermediate, Advanced, Graduate]
        tags:
          type: array
          items: { type: string }
...
    Community:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        banner: { type: string, nullable: true }
        color: { type: string }
        members: { type: integer }
        posts: { type: integer }
        tags:
          type: array
          items: { type: string }
        isJoined: { type: boolean }
        university: { type: string, nullable: true }
    CreateCommunityRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        banner: { type: string }
        color: { type: string }
        tags:
          type: array
          items: { type: string }
        university: { type: string }
...
    Conversation:
      type: object
      properties:
        id: { type: string }
        participants:
          type: array
          items: { type: string }
        isGroup: { type: boolean }
        groupName: { type: string, nullable: true }
        lastMessage: { type: string }
        lastTimestamp: { type: string, format: date-time }
        unread: { type: integer }
    CreateConversationRequest:
      type: object
      required: [participants]
      properties:
        participants:
          type: array
          items: { type: string }
        isGroup: { type: boolean }
        groupName: { type: string }
    UpdateConversationRequest:
      type: object
      properties:
        groupName: { type: string }
    Message:
      type: object
      properties:
        id: { type: string }
        senderId: { type: string }
        content: { type: string }
        timestamp: { type: string, format: date-time }
    CreateMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
...
    ConnectionStatusResponse:
      type: object
      properties:
        userId: { type: string }
        status:
          type: string
          enum: [none, pending, incoming, connected]
...
    PaginatedUsers:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/User" }
        total: { type: integer }
    PaginatedCourses:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Course" }
        total: { type: integer }
    PaginatedCommunities:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Community" }
        total: { type: integer }
...
    PaginatedConversations:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Conversation" }
        total: { type: integer }
    PaginatedMessages:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Message" }
        total: { type: integer }
...